day07 clean up done. All intcode exercises are being solved by a single intCode now. hvae to nwo review Day 09

+++ Day07/FeedbackAmplifiers/PermutationCalculator.cs
@@ -4,23 +4,23 @@ namespace FeedbackAmplifiers
 {
   public static class PermutationCalculator
   {
-    static List<List<int>> permutations;
+    static List<List<long>> permutations;
     static bool[] used;
 
-    public static List<List<int>> GetCombinations(int[] arr)
+    public static List<List<long>> GetCombinations(long[] arr)
     {
       used = new bool[arr.Length];
-      permutations = new List<List<int>>();
-      List<int> c = new List<int>();
+      permutations = new List<List<long>>();
+      List<long> c = new List<long>();
       GetPermutations(arr, 0, c);
       return permutations;
     }
 
-    private static void GetPermutations(int[] arr, int colindex, List<int> c)
+    private static void GetPermutations(long[] arr, int colindex, List<long> c)
     {
       if (colindex >= arr.Length)
       {
-        permutations.Add(new List<int>(c));
+        permutations.Add(new List<long>(c));
         return;
       }
       for (int i = 0; i < arr.Length; i++)
